feat:relacion user-todos dasahboard.server-todos por usuario: create Todo con usuario

+++ src/actions/todo.actions.ts
@@ -3,7 +3,8 @@
 import prisma from "@/lib/prisma";
 import { revalidatePath } from "next/cache";
 import { Todo } from "@prisma/client";
-
+import { actionGetSessionServer } from "./auth.actions";
+import { IActionsCrud } from "@/interface/actions.interface";
 
 const sleep = (seconds: number = 0) => {
     return new Promise((resolve) => {
@@ -33,26 +34,64 @@ export const actionToggleTodo = async (
     return todoUpdate;
 };
 
-export const actionInsertTodo = async (description: string) => {
-    try {
-        const todo = await prisma.todo.create({ data: { description } });
+export const actionInsertTodo = async (
+    description: string
+): Promise<IActionsCrud> => {
+    const resp: IActionsCrud = {
+        data: null,
+        success: false,
+        message: "",
+    };
 
+    try {
+        const session = await actionGetSessionServer();
+        if (!(session && session.user && !!session.user.id))
+            resp.message = "no-session";
+        else {
+            const newTodo = await prisma.todo.create({
+                data: {
+                    description,
+                    userId: session.user.id,
+                },
+            });
+            resp.data = newTodo;
+            resp.success = true;
+        }
         revalidatePath("dashboard/server-todos");
-        return todo;
     } catch (error) {
-        return { message: "Error al insertar" };
+        resp.message = `${error}`;
     }
+
+    return resp;
 };
 
-export const actionDeleteAllTodo = async (): Promise<number> => {
-    try {
-        const { count } = await prisma.todo.deleteMany({
-            where: { complete: true },
-        });
-
-        revalidatePath("dashboard/server-todos");
-        return count;
-    } catch (error) {
-        return 0;
+export const actionDeleteAllTodo = async (): Promise<IActionsCrud> => {
+    const resp:IActionsCrud = {
+        data: null,
+        success:false,
+        message:''
     }
+    
+    try {
+        const session = await actionGetSessionServer();
+        if (!(session && session.user && !!session.user.id)) 
+            resp.message = 'no-session'
+        else {
+            console.log('ID!!:', session.user.id)
+            const { count } = await prisma.todo.deleteMany({
+                where: { 
+                    userId: session.user.id,
+                    complete: true 
+                },
+            });
+            resp.data = count
+            resp.success = true
+    
+            revalidatePath("dashboard/server-todos");
+        }
+    } catch (error) {
+        resp.message = `${error}`;
+    }
+
+    return resp
 };
